src/requirements.txt
fastapi>=0.114.1
boto3>=1.35.18
uvicorn>=0.30.6
ddtrace>=2.12.2
PyJWT>=2.9.0
requests>=2.32.3
cryptography>=43.0.1
zipp==3.19.1